requirements.txt
streamlit
python-dotenv
pinecone-client
transformers
werkzeug
PyPDF2
openai
numpy==1.26.4
torch==2.3.1
torchvision==0.18.1
groq